Rename common module and fix drmcp tests
