+++ src/datarobot_genai/drmcp/core/dynamic_tools/deployment/controllers.py
@@ -16,11 +16,11 @@ import logging
 
 from fastmcp.tools.tool import Tool
 
-from datarobot_genai.drmcp.core.common import get_sdk_client
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.register import (
     register_tool_of_datarobot_deployment,
 )
 from datarobot_genai.drmcp.core.mcp_instance import mcp
+from datarobot_genai.drmcp.core.shared import get_sdk_client
 
 logger = logging.getLogger(__name__)
