+++ src/datarobot_genai/drmcp/core/dynamic_tools/deployment/metadata.py
@@ -18,12 +18,12 @@ from typing import Any
 import datarobot as dr
 from datarobot.utils import from_api
 
-from datarobot_genai.drmcp.core.common import get_api_client
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.adapters.base import MetadataBase
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.adapters.default import Metadata
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.adapters.drum import DrumMetadataAdapter
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.adapters.drum import DrumTargetType
 from datarobot_genai.drmcp.core.dynamic_tools.deployment.adapters.drum import is_drum
+from datarobot_genai.drmcp.core.shared import get_api_client
 
 logger = logging.getLogger(__name__)
