+++ src/datarobot_genai/drmcp/test_utils/mcp_utils_integration.py
@@ -22,7 +22,7 @@ from mcp import ClientSession
 from mcp.client.stdio import StdioServerParameters
 from mcp.client.stdio import stdio_client
 
-from .common import load_env
+from .shared import load_env
 
 load_env()
