+++ src/datarobot_genai/drmcp/tools/predictive/model.py
@@ -18,8 +18,8 @@ from typing import Any
 
 from datarobot.models.model import Model
 
-from datarobot_genai.drmcp.core.common import get_sdk_client
 from datarobot_genai.drmcp.core.mcp_instance import dr_mcp_tool
+from datarobot_genai.drmcp.core.shared import get_sdk_client
 
 logger = logging.getLogger(__name__)
