+++ tests/drmcp/unit/predictive_tools/test_predict_realtime.py
@@ -23,8 +23,8 @@ from unittest.mock import patch
 import pandas as pd
 import pytest
 
-from datarobot_genai.drmcp.core.common import MCPError
 from datarobot_genai.drmcp.core.constants import MAX_INLINE_SIZE
+from datarobot_genai.drmcp.core.shared import MCPError
 from datarobot_genai.drmcp.tools.predictive import predict_realtime
 from datarobot_genai.drmcp.tools.predictive.predict_realtime import make_output_settings
 from datarobot_genai.drmcp.tools.predictive.predict_realtime import predict_by_ai_catalog_rt
